% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{app_sys}
\alias{app_sys}
\title{Access files in the current app}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Character vector specifying directory and or file to
point to inside the current package.}
}
\description{
Access files in the current app
}
