% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{iraceStudioResources}
\alias{iraceStudioResources}
\title{Add external Resources to the Application}
\usage{
iraceStudioResources()
}
\description{
This function is internally used to add external
resources inside the Shiny application.
}
