% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_irace_studio}
\alias{run_irace_studio}
\title{Run Irace Studio}
\usage{
run_irace_studio(port = 4350, ...)
}
\arguments{
\item{port}{A port number that Irace Studio will listen on.}
}
\description{
Run Irace Studio
}
