% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{runIraceStudio}
\alias{runIraceStudio}
\title{Run Irace Studio}
\usage{
runIraceStudio(port = 4350)
}
\arguments{
\item{port}{A port number that Irace Studio will listen on.}
}
\description{
Run Irace Studio
}
